Split out get_axis_api to its own module
