+++ homeassistant/components/axis/hub/__init__.py
@@ -1,3 +1,4 @@
 """Internal functionality not part of HA infrastructure."""
 
-from .hub import AxisHub, get_axis_api  # noqa: F401
+from .api import get_axis_api  # noqa: F401
+from .hub import AxisHub  # noqa: F401
